backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.9.2
python-multipart==0.0.12
alembic==1.13.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
httpx==0.27.2
pyzbar==0.1.9
Pillow==10.4.0
